Actors are no longer generic over messages.
Actually now Actors receive a Box<Any> instead of a message.
This is done so that we can forward messages, have routers and servers.
Indeed if you want to pass messages to another machine, the actor in
charge of sending them will receive messages of any single possible type
and thus the previous model with genericity over message type is no
longer possible.